Remove podsd and -d from repods.
Reasoning: repods will now be for FULL refresh, and will not
run in the background, so devs will see logs and use crtl+c to
stop the containers. podsu simply runs docker-compose up to
start the containers. podsd is not longer necessary because
repods covers everything, and ctrl+c stops containers.